chore: Split `writer` module into several submodules
